pyproject.toml
# -*- coding: UTF-8 -*-
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# Package meta-data
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[project]
dynamic = ["version"]
name = "p4est"
description = "A library to manage a collection (a forest) of multiple connected adaptive quadtrees or octrees in parallel."
readme = "README"
license = { file = "COPYING" }

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# Package dependencies
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
requires-python = ">= 3.6.2"

dependencies = [
  "partis-pyproj >= 0.0.4",
  "partis-schema >= 0.0.4",
  'numpy >= 1.18.5',
  "mpi4py >= 3.1.3",
  "lxml >= 4.2.5",
  "pyevtk >= 1.1",
  "vtk >= 8.1.2" ]

[[project.authors]]
email = "p4est@ins.uni-bonn.de"

[[project.authors]]
name = "Carsten Burstedde"
email = "burstedde@ins.uni-bonn.de"

[[project.authors]]
name = "Lucas C. Wilcox"
email = "me@lucaswilcox.com"

[[project.authors]]
name = "Tobin Isaac"
email = "tisaac@uchicago.edu"


#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[project.optional-dependencies]

doc = [
  'partis-utils[sphinx] >= 0.1.0' ]

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# Build and distribution configuration
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[build-system]
requires = [
  "partis-pyproj[meson] >= 0.0.6",
  "cython ~= 3.0a",
  # "cython >= 0.29.18",
  "numpy==1.16.5; python_version=='3.6'",
  "numpy==1.16.5; python_version=='3.7'",
  "numpy==1.18.5; python_version=='3.8'",
  "numpy==1.19.3; python_version=='3.9'",
  "numpy==1.21.4; python_version=='3.10'",
  "numpy; python_version>='3.11'",
  "mpi4py >= 3.1.3" ]

build-backend = "partis.pyproj.backend"

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[tool.pyproj.config]
# build with debugging
debug = false

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[tool.pyproj.prep]
# updates version
entry = "pkgaux:prep"

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[tool.pyproj.dist]
ignore = [
  '**/__pycache__',
  '**/*.py[cod]',
  '*.egg-info',
  '**/.git',
  '**/.github',
  '**/.gitignore',
  '**/.gitmodules' ]

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[tool.pyproj.dist.source]
ignore = [
  '**/*.so' ]

copy = [
  'cmake',
  'doc',
  'example',
  'pkgaux',
  'pypkg',
  'src',
  'subprojects/sc/cmake',
  'subprojects/sc/doc',
  'subprojects/sc/example',
  'subprojects/sc/iniparser',
  'subprojects/sc/libb64',
  'subprojects/sc/src',
  'subprojects/sc/test',
  'subprojects/sc/AUTHORS',
  'subprojects/sc/COPYING',
  'subprojects/sc/VERSION',
  'subprojects/sc/BUILDNUMBER.in',
  'subprojects/sc/meson_options.txt',
  'subprojects/sc/meson.build',
  'test',
  'pyproject.toml',
  'README',
  'COPYING',
  'VERSION',
  'BUILDNUMBER.in',
  'meson_options.txt',
  'meson.build',
  'build_requirements.txt']

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
[tool.pyproj.dist.binary.prep]
# updates build number and tags
entry = "pkgaux:dist_binary_prep"

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# Configuration of the meson build system for p4est
[[tool.pyproj.targets]]
entry = 'partis.pyproj.builder:meson'

# NOTE: for developement, setting 'build_clean = false' makes re-compiling faster,
# since it can skip repeating some steps, but be cautious of large changes, or
# different 'options', breaking an existing build config (delete the build directory).
build_clean = false
prefix = 'build'
build_dir = 'build/tmp'

setup_args = [
  '--debug',
  '--warnlevel', '0',
  '--libdir', 'p4est/core/lib',
  '--includedir', 'p4est/core/include' ]

compile_args = [
  '-j', '-1' ]

#...............................................................................
# p4est build options
[tool.pyproj.targets.options]

# Compiles the Python modules
pypkg = true

# always build against MPI
mpi = 'enabled'

# one of {'trace', 'debug', 'verbose', 'info', 'warning', 'error'}
log_priority = 'trace'

# Compile with profiling code
profile = false

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# binary distribution installed paths

[[tool.pyproj.dist.binary.purelib.copy]]
src = 'pypkg/p4est'
dst = 'p4est'
ignore = 'core'

[[tool.pyproj.dist.binary.purelib.copy]]
src = 'build/p4est'
dst = 'p4est'
glob = '**/*.py'

[[tool.pyproj.dist.binary.platlib.copy]]
src = 'build/p4est'
dst = 'p4est'
ignore = '**/*.py'

